Move some classes around a bit.
Wire protocol stuff into `baaahs.proto`.
Fakey simulation stuff into `baaahs.sim`.
