+++ src/commonMain/kotlin/baaahs/Show.kt
@@ -1,9 +1,9 @@
 package baaahs
 
-abstract class ShowMeta(val name: String) {
-    abstract fun createShow(sheepModel: SheepModel, showRunner: ShowRunner): Show
-}
-
 interface Show {
     fun nextFrame()
+
+    abstract class MetaData(val name: String) {
+        abstract fun createShow(sheepModel: SheepModel, showRunner: ShowRunner): Show
+    }
 }
